backend/requirements.txt
Django>=4.2.5,<5
psycopg2-binary==2.9.6
djangorestframework>=3.12,<4
django-cors-headers>=3.13.0,<4
cloudinary>=1.24.0,<2
gunicorn>=20.0.4,<21
python-dotenv>=0.19.0,<1
whitenoise>=5.2.0,<6
djangorestframework-simplejwt==5.2.2
dj-database-url==2.2.0
django-filters==2.0.0
drf-spectacular==0.26.3